requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.15.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
